Use a BigInt to store 7 * 24 * 6 bit
