+++ .eslintrc.json
@@ -19,6 +19,7 @@
         "react"
     ],
     "rules": {
-        "no-use-before-define": "off"
+        "no-use-before-define": "off",
+        "no-bitwise": "off"
     }
 }
